refactor: split test files
